Auto-Files-Organization
In a folder or in a drive we have different files with different extensions like images 
having extension [.jpg, .png, .jpeg], videos [extension .mp4, .mkv, .flv, .avi, ], pdfs, 
docx file and so many more. That's a bit tough to manage all these files. so, I have done a program for these auto files 
an organization that manages the files and with different extensions and copy the files and kept the same group of files like all images 
into an image folder and all videos into a videos folder and so on...